homepage fix
